Add files via upload
This project focuses on the Exploratory Data Analysis (EDA) of a YouTube dataset that provides key insights into some of the top YouTube channels and their metrics. The main goal of this analysis is to understand the distribution and characteristics of top YouTubers based on several factors such as the number of subscribers, total video views, and the categories they create content in. By analyzing these factors, we can better understand the popularity trends on YouTube and the key drivers of channel success.